+++ src/mol-repr/structure/visual/molecular-surface-wireframe.ts
@@ -1,7 +1,8 @@
 /**
- * Copyright (c) 2019-2025 mol* contributors, licensed under MIT, See LICENSE file for more info.
+ * Copyright (c) 2019-2026 mol* contributors, licensed under MIT, See LICENSE file for more info.
  *
  * @author Alexander Rose <alexander.rose@weirdbyte.de>
+ * @author Gianluca Tomasello <giagitom@gmail.com>
  */
 
 import { ParamDefinition as PD } from '../../../mol-util/param-definition';
@@ -10,13 +11,14 @@ import { VisualContext } from '../../visual';
 import { Unit, Structure } from '../../../mol-model/structure';
 import { Theme } from '../../../mol-theme/theme';
 import { Lines } from '../../../mol-geo/geometry/lines/lines';
-import { CommonMolecularSurfaceCalculationParams, computeUnitMolecularSurface } from './util/molecular-surface';
+import { CommonMolecularSurfaceCalculationParams, computeUnitMolecularSurface, computeStructureMolecularSurface } from './util/molecular-surface';
 import { computeMarchingCubesLines } from '../../../mol-geo/util/marching-cubes/algorithm';
-import { ElementIterator, getElementLoci, eachElement } from './util/element';
+import { ElementIterator, getElementLoci, eachElement, getSerialElementLoci, eachSerialElement } from './util/element';
 import { VisualUpdateState } from '../../util';
 import { CommonSurfaceParams } from './util/common';
 import { Sphere3D } from '../../../mol-math/geometry';
 import { Tensor } from '../../../mol-math/linear-algebra/tensor';
+import { ComplexLinesVisual, ComplexVisual } from '../complex-visual';
 
 export const MolecularSurfaceWireframeParams = {
     ...UnitsLinesParams,
@@ -54,6 +56,47 @@ export function MolecularSurfaceWireframeVisual(materialId: number): UnitsVisual
         createLocationIterator: ElementIterator.fromGroup,
         getLoci: getElementLoci,
         eachLocation: eachElement,
+        setUpdateState: (state: VisualUpdateState, newProps: PD.Values<MolecularSurfaceWireframeParams>, currentProps: PD.Values<MolecularSurfaceWireframeParams>) => {
+            state.createGeometry = (
+                newProps.resolution !== currentProps.resolution ||
+                newProps.probeRadius !== currentProps.probeRadius ||
+                newProps.probePositions !== currentProps.probePositions ||
+                newProps.ignoreHydrogens !== currentProps.ignoreHydrogens ||
+                newProps.ignoreHydrogensVariant !== currentProps.ignoreHydrogensVariant ||
+                newProps.includeParent !== currentProps.includeParent ||
+                newProps.floodfill !== currentProps.floodfill
+            );
+        }
+    }, materialId);
+}
+
+//
+
+async function createStructureMolecularSurfaceWireframe(ctx: VisualContext, structure: Structure, theme: Theme, props: MolecularSurfaceWireframeProps, lines?: Lines): Promise<Lines> {
+    const { transform, field, idField, maxRadius } = await computeStructureMolecularSurface(structure, theme.size, props).runInContext(ctx.runtime);
+
+    const params = {
+        isoLevel: props.probeRadius,
+        scalarField: props.floodfill !== 'off' ? Tensor.createFloodfilled(field, props.probeRadius, props.floodfill) : field,
+        idField
+    };
+    const wireframe = await computeMarchingCubesLines(params, lines).runAsChild(ctx.runtime);
+
+    Lines.transform(wireframe, transform);
+
+    const sphere = Sphere3D.expand(Sphere3D(), structure.boundary.sphere, maxRadius);
+    wireframe.setBoundingSphere(sphere);
+
+    return wireframe;
+}
+
+export function StructureMolecularSurfaceWireframeVisual(materialId: number): ComplexVisual<MolecularSurfaceWireframeParams> {
+    return ComplexLinesVisual<MolecularSurfaceWireframeParams>({
+        defaultProps: PD.getDefaultValues(MolecularSurfaceWireframeParams),
+        createGeometry: createStructureMolecularSurfaceWireframe,
+        createLocationIterator: ElementIterator.fromStructure,
+        getLoci: getSerialElementLoci,
+        eachLocation: eachSerialElement,
         setUpdateState: (state: VisualUpdateState, newProps: PD.Values<MolecularSurfaceWireframeParams>, currentProps: PD.Values<MolecularSurfaceWireframeParams>) => {
             state.createGeometry = (
                 newProps.resolution !== currentProps.resolution ||